Implement path zipping
With the introduction of the custom Point struct, and its subsequent adoption in all the internal path processing algorithms, the necessity arose to be able to convert the seperate x array and y array given by the frontend javascript into the necessary Point vector. This new function solves that problem with a time complexity of O(n), by simply iterating over the two elements and building a temporary vector, which is then returned when the function is done.