Refactor some more
It was still a little cluttered and not very open to additions, so I refactored some more. I added a Section type, which is just an alias for u8, once again to improve readability. All of the changes were internal structural changes, and none of the actual implementations or algorithms were changed, and the module is still accessed the exact same externally.