Implement direction derivation
I added the ability to determine the direction between two points, then used it to add the ability to reduce a path down to its directions. In order to facilitate this, I also added a new variant on the Direction enum, `Still`, which indicates that there is no movement whatsoever.