Implement corner detection
Sorry for the long time this took, I've been dealing with finals and this algorithm was really hard to right cause i was measuring angles over the wrong places cause I'm an idiot. Anyway, it's written now and it's still really bad and needs improvements. In order to facilitate the new algorithm I added some helper functions, and modified the test code as well as the frontend.